Footerの作成と、それに伴う画面遷移の対応
