+++ lib/list_chat.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+
+class ListChat extends StatelessWidget {
+  const ListChat({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      child: Center(child: Text('お知らせ')),
+      color: Colors.red[200],
+    );
+  }
+}
